test: merge rename-registry tests into one file
test_field_renames.py (Pydantic field rename registry, dual-read,
deprecation warnings) and test_df_column_renames.py (the DataFrame
column registry that mirrors it, read_df_column, Rust table alignment)
covered the two halves of the same rename surface. They now live in
test_renames.py with bodies unchanged; the DF half's colliding
TestRegistryIntegrity class is renamed TestDfRegistryIntegrity.
136 collected items before and after. The cross-reference comment in
test_validation.py is updated to the new file name.
